+++ examples/pinMode/pinMode.c
@@ -4,12 +4,12 @@
 
 #define BLINK_DELAY_MS 50
 
-int main (void)
+uint16_t main (void)
 {
     uint8_t UNO_pin = 13;
     on_led();
     pinMode(UNO_pin, OUTPUT);
-    for (int i=0;i<40;i++) {
+    for (uint8_t i=0;i<40;i++) {
         digitalWrite(UNO_pin, HIGH);
         _delay_ms(BLINK_DELAY_MS);
         digitalWrite(UNO_pin, LOW);
